week 2/requirements.txt
fastapi
uvicorn
ollama
Python-dotenv
requests

langchain
langchain-ollama
langchain-chroma
pandas